+++ app/Http/Controllers/Login/LoginController.php
@@ -7,5 +7,14 @@ use Illuminate\Http\Request;
 
 class LoginController extends Controller
 {
+    public function disclaimer()
+    {
+        return view('auth.disclaimer');
+    }
 
+    public function privacy()
+    { // Legal
+       // Privacy Policy
+        return view('auth.privacy');
+    }
 }